Adds the Detailed list of Klimadao retirements on the token page
