+++ carbonmark-data/components/cards/retirementTrends/KlimaDAORetirementsByTokenListCard/index.tsx
@@ -1,0 +1,19 @@
+import { t } from "@lingui/macro";
+import ChartCard, { CardProps } from "components/cards/ChartCard";
+import DataTable from "components/charts/helpers/DataTable";
+
+export default function KlimaDAORetirementsByTokenListCard(props: CardProps) {
+  const chart = (
+    /* @ts-expect-error async Server component */
+    <DataTable configurationKey="KlimaRetirementsByTokenList"></DataTable>
+  );
+
+  return (
+    <ChartCard
+      {...props}
+      title={t`Detailed list of KlimaDAO retirements`}
+      isColumnCard={true}
+      chart={chart}
+    />
+  );
+}
