+++ carbonmark-data/components/charts/helpers/DataTable/configurations/index.ts
@@ -1,9 +1,11 @@
 import KlimaRetirementsByPoolListConfiguration from "./KlimaRetirementsByPoolListConfiguration";
 import KlimaRetirementsByPoolSummaryConfiguration from "./KlimaRetirementsByPoolSummaryConfiguration";
+import KlimaRetirementsByTokenListConfiguration from "./KlimaRetirementsByTokenListConfiguration";
 const configurations = {
   KlimaRetirementsByPoolSummary:
     new KlimaRetirementsByPoolSummaryConfiguration(),
   KlimaRetirementsByPoolList: new KlimaRetirementsByPoolListConfiguration(),
+  KlimaRetirementsByTokenList: new KlimaRetirementsByTokenListConfiguration(),
 };
 export type ConfigurationKey = keyof typeof configurations;
 function getConfiguration(key: ConfigurationKey) {
